requirements.txt
# FastAPI & Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
supabase==2.3.4
sqlalchemy==2.0.25

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Utils
python-dotenv==1.0.1
httpx==0.26.0

# File handling
aiofiles==23.2.1
python-magic-bin==0.4.14;sys_platform=='win32'
python-magic==0.4.27;sys_platform!='win32'

